Move memory sync test to test_unit.py
